feat: pathfinding with A* and JPS
Added a general purpose Heap data structure using
a flat array internally for better performance.

Added A* and JPS pathfinding algorithms, for
pathfinding across a 2D grid.

Extended mouse presses to include when a button
was just clicked.

Added Vector2i struct with basic operators.

Separated rendering of tiles to separate function
to set specific debugging colors on the tiles.

Reworked how player position is used, now the
position is kept "simple" and pathing moves
the player from tile to tile, and rendering
offsets the position so it centers it on a tile
with the correct "feet" position.

Changed Player and PLAYER_STATE to Entity and
ENTITY_STATE, and added is_walkable, is_hovered
and is_path to Tile to aid debugging.