refactor: .bss GameState data
Changed entity pathfinding to have a finite
list of paths it navigates, in order to easily
store the entire GameState in contiguous memory.

This also cuts the path of at max path length in
an intuitive way by cutting of the last paths to
fit the max path length. This changed also
included traversing the path in reverse, instead
of reversing the path from the pathfinding algo.

Updated main to store 256 GameStates in .bss
as well as 256 InputStates, and for them to loop
around when updating. The 'next' GameState is
just simulated from the previous state + inputs.
This is necessary to easily integrate server
reconciliation for multiplayer in the future.

Updated input handling to accumulate and take
into account the input of the previous tick.

Moved GameState view_pos and view_zoom
to a View struct.

Updated render interpolation to interpolate
between two GameStates, instead of storing a
previous position on the Entity struct.