feat: networking stub continuation
Too many individual changes, but to summarize;

Added more network messages, and the connection
flow was updated so that a client connects,
server sends some init data, and then broadcasts
a new player to all connected clients.

Added a separate main-server.jai that deals with
server-only logic. Also stripped out App stuff
to app.jai for shrinking my brain-context.

Added ActionRequest and ActionResult to be able
to simply perform operations on the players
from the server, without the server doing any
playing. This concept needs further fleshing
out to potentially strip away the entire
input_state into simulate(), and instead receive
a list of ActionRequests converted from the
accumulated inputs.

Got a working stub of movement synchronizing
from client to server, which includes fast
forwarding the state on the server from the time
the action occurred to where the server is
currently.