feat: networking stub final
Moved some shared functionality and constants to app.jai and renamed
it to infrastructure.jai

Added a list of actions to GameState and added some function for
adding and removing actions from a game state, as well as a helper
function for copying game state while keeping actions intact.

Added uuid to actions to refer to a specific client, instead of
supplying an entity into perform_action_request.

Updated entity to store path indices instead of path positions, and
added an interpolated_position and interpolating boolean.
Also updated renderer to draw interpolated pos when interpolating.

Added PlayerMoveResponseMessage to send from server to client in order
to correct and reconcile the path of a specific client. This is also
used with the interpolated_position on the entity to simulate
smoother movement of other clients.

Added convert_input_to_actions to start moving away from reading
external input_state in simulate().