refactor: move view to app and simplify input
Moved View from game.jai to view.jai, and added it to AppState along
with updating it outside of simulate(), so that view updates happens
as fast as possible.

Since the view is used for calculations within the game state though,
the view in the GameState is kept, and is just updated from the app
view before simulating.

Move back InputState into the GameState since it makes more sense,
any actions of transient states are already converted to actions and
flipped on the accumulated input so it doesn't actually go into the
game states' input state.

Also simplified copy_game_state_keeping_actions and renamed it to
copy_game_state.