refactor: general networking improvements
Moved message processing to a separate function to easier
decouple network vs single player code.

Converted space input to an action to further separate network code.
It gets handled by default in simulate and is sent to
server if multiplayer. To get around double actions without adding
a reconciliation step to it, an ignore flag was added to actions.

Separated out network message data to a struct, which made creation
a lot simpler since it could be assigned directly to a struct literal.

Simplified vector2i serialization by adding it
as an overload to serialize_struct.

Changed path in PlayerMoveResponse to a fixed length array instead of
dynamically allocating enough space, which proved simpler to handle
and less complicated to implement. This may change in the future
if a proper arena memory is used for network messages instead.

Renamed game_state_index on message to simply state_index, it didn't
need to be that verbose.

Updated renderer to only interpolate if there was a previous player.