feat: world refactor and area generation
Separated common.jai into clock.jai, vector.jai and common.jai under
src/common, and added rectangle.jai with Rect and IntRect structs.

Removed SDL_Rect from Sprite and replaced it with IntRect
and a conversion function.

Moved all tile specific stuff into a new World struct, and ripped
it out from GameState. This was then made into a single WORLD var in
infrastructure, since it doesn't make sense to have copy of the world
for each game state.

Added Area struct containing a rectangle and an entrance, and a
function to generate a random Area on the map.

The entrance is comprised of 3 points on a random side of the
rectangle. If the entrance is on a side directly adjacent to the edge
of the world, we randomize a new side until a "good" side is found.