fix: remove client interpolation stutter
This was fixed by moving the fetch of the old position to before
the reconciliation occurred. The main issue here was that when the
msg state index was the previous state, the position fetched was
different after reconciliation than before.

Updated how the interpolation is done in simulate by using the
same functionality as the normal movement is done. This means that
the player on the screen will be delayed and never up-to-date, but
this will (potentially) simplify future reconciliation.

Also added SDL_Delay(1) to not hog CPU.