feat: claim an area over the wire
Added back an ActionResult struct which is stored on ActionRequest.
An action populates the result appropriately, so that we can record
what the action did.

Pulled out send_message to two functions in network.jai, depending
on if we're sending from client or server.

Also pulled out network_manager_update to two functions with
overloads for server or client.

Finally, I updated Yojimbo to latest, which slimmed down the message
structs a bit.