refactor: generic network improvements
Moved all messages from server to client to be handled outside
of game, since they just replace "reality" for the specific state.

Moved all client message processing to client.jai, and separated
all messages into separate functions instead of everything handled
in the switch-case in process_messages.

The same changes was done for the server, which was moved
to server.jai.